+++ packages/guardian/examples/aggregate-boundary/domain/aggregates/product/Product.ts
@@ -1,0 +1,7 @@
+export class Product {
+    public price: number
+
+    constructor(price: number) {
+        this.price = price
+    }
+}